updated string library - fixed todos

+++ kernel/types/string.h
@@ -7,7 +7,7 @@
 char* malloc_str(const uint64_t size);
 
 uint64_t str_length(const char* str);
-char* str_repeat(const char* text, int amount);
+char* str_repeat(const char* text, uint64_t amount);
  
 bool str_equal(const char* a, const char* b);
 
@@ -16,7 +16,7 @@ void str_add(char* target, const char* value);
 char* str_combine(const char* a, const char* b);
 
 bool str_contains(const char* text, const char* searchText);
-int str_count(const char* text, const char* searchText);
+uint64_t str_count(const char* text, const char* searchText);
 bool str_starts_with(const char* text, const char* searchText);
 bool str_ends_with(const char* text, const char* searchText);
 //void str_replace(char* text, const char* toBeReplacedText, const char* replacementText);
